+++ src/packages/analytics/index.ts
@@ -7,7 +7,7 @@ const endpoint = {
 };
 
 interface AnalyticsOptions {
-  env?: 'local' | 'stage' | 'prod';
+  env: 'local' | 'stage' | 'prod';
 }
 
 /**
@@ -21,15 +21,8 @@ interface AnalyticsOptions {
  * @param {Object} [options] Analytics options
  * @param {string} [options.env] 'prod' (default) or 'stage'
  */
-export default function report(
-  evt: AnalyticsEvent,
-  userOptions: AnalyticsOptions = { env: 'prod' }
-) {
-  const options: AnalyticsOptions = {
-    env: 'prod',
-    ...(userOptions || {}),
-  };
-  const url = endpoint[options.env || 'prod'];
+export default function report(evt: AnalyticsEvent, options: AnalyticsOptions) {
+  const url = endpoint[options.env];
   return fetch(url, {
     method: 'POST',
     body: JSON.stringify({