+++ src/packages/analytics/types.ts
@@ -20,26 +20,26 @@ export type EventTypes =
         initialRender: number;
         renderWithData: number;
         rttGraphql: number;
-        time: number;
+        duration: number;
       };
     }
   | {
       name: 'render_with_data';
       properties: {
         rttGraphql: number;
-        time: number;
+        duration: number;
       };
     }
   | {
       name: 'rtt_graphql';
       properties: {
-        time: number;
+        duration: number;
       };
     }
   | {
       name: 'token_received';
       properties: {
-        time: number;
+        duration: number;
       };
     };