+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Added `metrics` property to `<manifold-connection>` to opt-in for metric collection (#724)
+- Added performance monitoring of GraphQL endpoints (#726)
 
 ## [0.6.5] - 2019-11-13