+++ CHANGELOG.md
@@ -12,6 +12,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `metrics` property to `<manifold-connection>` to opt-in for metric collection (#724)
 - Added performance monitoring of GraphQL endpoints (#726)
 - Added performance monitoring of oauth token handshake (#730)
+- Added JS minification (#734)
 
 ### Removed