+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.0] - 2019-11-19
 
 ### Added
 
@@ -429,6 +429,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Changed the `manifold-resource-credentials` component to use the standalone `manifold-credentials`
   component.
 
+[0.7.0]: https://github.com/manifoldco/ui/compare/v0.6.5...v0.7.0
 [0.6.5]: https://github.com/manifoldco/ui/compare/v0.6.4...v0.6.5
 [0.6.4]: https://github.com/manifoldco/ui/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/manifoldco/ui/compare/v0.6.2...v0.6.3